Add user interface
Implemented a new user interface for the "Rock-Paper-Scissors" game,
allowing players to interact with the game through a graphical
interface. Added the ability to make selections by clicking buttons,
and the display of match results.
Additionally, included a restart button that utilizes DOM manipulations
to reset the game for another round.